fix: move NFTAsset type to global level
