+++ packages/client/lib/Nft.ts
@@ -1,4 +1,4 @@
-import { AddressType, BigNumber, FeeType, Transaction } from '@chainify/types';
+import { AddressType, BigNumber, FeeType, NFTAsset, Transaction } from '@chainify/types';
 import Wallet from './Wallet';
 
 export default abstract class Nft<T, S> {
@@ -33,5 +33,5 @@ export default abstract class Nft<T, S> {
 
     public abstract isApprovedForAll(contract: AddressType, operator: AddressType): Promise<boolean>;
 
-    public abstract fetch(): void;
+    public abstract fetch(): Promise<NFTAsset[]>;
 }
