+++ packages/evm/lib/nft/EvmNftProvider.ts
@@ -1,11 +1,11 @@
 import { Nft } from '@chainify/client';
 import { UnsupportedMethodError } from '@chainify/errors';
-import { AddressType, BigNumber, FeeType, Transaction } from '@chainify/types';
+import { AddressType, BigNumber, FeeType, NFTAsset, Transaction } from '@chainify/types';
 import { Signer } from '@ethersproject/abstract-signer';
 import { AddressZero } from '@ethersproject/constants';
 import { BaseProvider } from '@ethersproject/providers';
 import { ERC1155, ERC1155__factory, ERC721, ERC721__factory } from '../typechain';
-import { EthersPopulatedTransaction, EthersTransactionResponse, NFTAsset, NftTypes } from '../types';
+import { EthersPopulatedTransaction, EthersTransactionResponse, NftTypes } from '../types';
 import { toEthereumTxRequest } from '../utils';
 import { EvmBaseWalletProvider } from '../wallet/EvmBaseWalletProvider';
